refac[Form] implementing the initial Form
One of the challenge requisits is that should be exist a form input  to
type the player name and button for router to game. Now we have a form
that works fine, and redirect fine too, but lacks style, let keep it as
it is now and focus in the game instead.