config[eslint] Set eslint to not interfere in methods
Eslint was firing problems when methods didnt use the keyword 'this' but
it looks like a bug in eslint or ahi lake of understanding the rule. we
set the rule to 'off' since it was cutting productivity trying to fix
the problem. In the future we can refactor to handle this issue if necessary