refact[movement] Implements a keyboard listener
Currencty we have a problem firing the event.preventDefault on player
movements, since we have a form in the main page, and the keyboard action
are not firing, so we are desacoplating the movement into a oberserver
pattern where the subject is the keypressed event. So we also can solve
the input bug.